backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
et-xmlfile==2.0.0
openpyxl==3.1.5
Pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
reportlab==4.4.0
sqlparse==0.5.3
xlrd==2.0.1
# mysqlclient==2.2.7 # Uncomment for production with MySQL 